refactor: rename rule to catchDieToOrDie
