+++ .changeset/add-catch-die-to-or-die.md
