+++ docs/rules/catch-die-to-or-die.md
@@ -1,6 +1,6 @@
 <!-- This file is generated by internal/rules/rules_json_test.go. Do not edit it manually. -->
 
-# `catchAllDieToOrDie`
+# `catchDieToOrDie`
 
 Suggests using Effect.orDie instead of Effect.catch or Effect.catchAll with an identity-forwarding Effect.die handler
 
@@ -11,8 +11,8 @@ Suggests using Effect.orDie instead of Effect.catch or Effect.catchAll with an i
 | Fixable | Yes |
 | Effect versions | v3, v4 |
 | Diagnostic codes | `TS377115` |
-| Language Service name | `catchAllDieToOrDie` |
-| Oxlint name | `effecttsgo/catch-all-die-to-or-die` |
+| Language Service name | `catchDieToOrDie` |
+| Oxlint name | `effecttsgo/catch-die-to-or-die` |
 
 ## Preview
 
@@ -24,7 +24,7 @@ declare const readConfig: Effect.Effect<string, Error>
 export const program = readConfig.pipe(
   Effect.catch((error) => Effect.die(error))
 /**
-  ^^^^^^^^^^^^ effecttsgo(catch-all-die-to-or-die): `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catch` with an identity-forwarding `Effect.die` handler.
+  ^^^^^^^^^^^^ effecttsgo(catch-die-to-or-die): `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catch` with an identity-forwarding `Effect.die` handler.
 */
 )
 ```
@@ -41,7 +41,7 @@ See the [Language Service setup guide](../../README.md#installation) for install
       {
         "name": "@effect/language-service",
         "diagnosticSeverity": {
-          "catchAllDieToOrDie": "warning"
+          "catchDieToOrDie": "warning"
         }
       }
     ]
@@ -61,7 +61,7 @@ See the [Oxlint setup guide](../README.md#oxlint-setup) for installation and pat
   },
   "plugins": ["effecttsgo"],
   "rules": {
-    "effecttsgo/catch-all-die-to-or-die": "warn"
+    "effecttsgo/catch-die-to-or-die": "warn"
   }
 }
 ```
