--- testdata/baselines/reference/effect-v3/catchAllDieToOrDie.errors.txt
@@ -1,89 +1,0 @@
-=== Metadata ===
-Effect version: 3.19.19
-
-/.src/catchAllDieToOrDie.ts(10,41): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(14,3): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(17,3): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(22,3): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(28,44): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(29,26): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(30,25): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-/.src/catchAllDieToOrDie.ts(35,4): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-
-
-==== /.src/catchAllDieToOrDie.ts (8 errors) ====
-    // @effect-v3
-    // @effect-diagnostics *:off
-    // @effect-diagnostics catchAllDieToOrDie:warning
-    import * as Effect from "effect/Effect"
-    import { pipe } from "effect/Function"
-    
-    declare const failable: Effect.Effect<number, Error>
-    
-    // Should trigger: direct API reference.
-    export const pipeDirect = failable.pipe(Effect.catchAll(Effect.die))
-                                            ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    
-    // Should trigger: concise and block identity-forwarding functions.
-    export const pipeArrow = failable.pipe(
-      Effect.catchAll((error) => Effect.die(error))
-      ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    )
-    export const pipeArrowBlock = failable.pipe(
-      Effect.catchAll((error) => {
-      ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-        return Effect.die(error)
-      })
-    )
-    export const pipeFunction = failable.pipe(
-      Effect.catchAll(function(error) {
-      ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-        return Effect.die(error)
-      })
-    )
-    
-    // Should trigger: Function.pipe and both dual application forms.
-    export const functionPipe = pipe(failable, Effect.catchAll(Effect.die))
-                                               ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    export const dataFirst = Effect.catchAll(failable, Effect.die)
-                             ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    export const dataLast = Effect.catchAll(Effect.die)(failable)
-                            ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    
-    // Should trigger: Effect.fn trailing transformations are piping flows too.
-    export const effectFn = Effect.fn(function*() {
-      return yield* failable
-    }, Effect.catchAll(Effect.die))
-       ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    
-    // Should not trigger: the failure is transformed before becoming a defect.
-    export const wrapped = failable.pipe(
-      Effect.catchAll((error) => Effect.die(new Error(String(error))))
-    )
-    
-    // Should not trigger: this is not identity forwarding.
-    export const differentValue = failable.pipe(
-      Effect.catchAll((_error) => Effect.die("different"))
-    )
-    
-    // Should not trigger: Effect.orDie would itself be redundant.
-    export const unfailable = Effect.succeed(1).pipe(
-      Effect.catchAll(Effect.die)
-    )
-    
-    // Should not trigger: a local API with the same member names.
-    const LocalEffect = {
-      catchAll: (_handler: (error: Error) => Effect.Effect<never>) =>
-        <A, E, R>(self: Effect.Effect<A, E, R>) => self,
-      die: Effect.die
-    }
-    export const localApi = failable.pipe(LocalEffect.catchAll(LocalEffect.die))
-
