--- testdata/baselines/reference/effect-v3/catchAllDieToOrDie.layers.txt
@@ -1,1 +1,0 @@
-==== /.src/catchAllDieToOrDie.ts (0 layer exports) ====
