--- testdata/baselines/reference/effect-v3/catchAllDieToOrDie_preview.errors.txt
@@ -1,20 +1,0 @@
-=== Metadata ===
-Effect version: 3.19.19
-
-/.src/catchAllDieToOrDie_preview.ts(9,3): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-
-
-==== /.src/catchAllDieToOrDie_preview.ts (1 errors) ====
-    // @effect-v3
-    // @effect-diagnostics *:off
-    // @effect-diagnostics catchAllDieToOrDie:warning
-    import * as Effect from "effect/Effect"
-    
-    declare const readConfig: Effect.Effect<string, Error>
-    
-    export const program = readConfig.pipe(
-      Effect.catchAll((error) => Effect.die(error))
-      ~~~~~~~~~~~~~~~
-!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catchAll` with an identity-forwarding `Effect.die` handler. effect(catchAllDieToOrDie)
-    )
-
