--- testdata/baselines/reference/effect-v3/catchAllDieToOrDie_preview.flows.txt
@@ -1,1 +1,0 @@
-/.src/catchAllDieToOrDie_preview.ts -> catchAllDieToOrDie_preview.flows.catchAllDieToOrDie_preview.mermaid
