--- testdata/baselines/reference/effect-v3/catchAllDieToOrDie_preview.layers.txt
@@ -1,1 +1,0 @@
-==== /.src/catchAllDieToOrDie_preview.ts (0 layer exports) ====
