+++ testdata/baselines/reference/effect-v3/catchDieToOrDie.flows.catchDieToOrDie.mermaid
