+++ testdata/baselines/reference/effect-v3/catchDieToOrDie.flows.txt
@@ -1,0 +1,1 @@
+/.src/catchDieToOrDie.ts -> catchDieToOrDie.flows.catchDieToOrDie.mermaid
