+++ testdata/baselines/reference/effect-v3/catchDieToOrDie.layers.txt
@@ -1,0 +1,1 @@
+==== /.src/catchDieToOrDie.ts (0 layer exports) ====
