+++ testdata/baselines/reference/effect-v3/catchDieToOrDie.pipings.txt
@@ -1,4 +1,4 @@
-==== /.src/catchAllDieToOrDie.ts (18 flows) ====
+==== /.src/catchDieToOrDie.ts (18 flows) ====
 
 === Piping Flow ===
 Location: 10:26 - 10:69
