+++ testdata/baselines/reference/effect-v3/catchDieToOrDie_preview.flows.txt
@@ -1,0 +1,1 @@
+/.src/catchDieToOrDie_preview.ts -> catchDieToOrDie_preview.flows.catchDieToOrDie_preview.mermaid
