+++ testdata/baselines/reference/effect-v3/catchDieToOrDie_preview.layers.txt
@@ -1,0 +1,1 @@
+==== /.src/catchDieToOrDie_preview.ts (0 layer exports) ====
