--- testdata/baselines/reference/effect-v4/catchAllDieToOrDie.flows.txt
@@ -1,1 +1,0 @@
-/.src/catchAllDieToOrDie.ts -> catchAllDieToOrDie.flows.catchAllDieToOrDie.mermaid
