--- testdata/baselines/reference/effect-v4/catchAllDieToOrDie.layers.txt
@@ -1,1 +1,0 @@
-==== /.src/catchAllDieToOrDie.ts (0 layer exports) ====
