+++ testdata/baselines/reference/effect-v4/catchDieToOrDie.flows.catchDieToOrDie.mermaid
