+++ testdata/baselines/reference/effect-v4/catchDieToOrDie.pipings.txt
@@ -1,4 +1,4 @@
-==== /.src/catchAllDieToOrDie.ts (19 flows) ====
+==== /.src/catchDieToOrDie.ts (19 flows) ====
 
 === Piping Flow ===
 Location: 9:26 - 9:66
