+++ testdata/baselines/reference/effect-v4/catchDieToOrDie_preview.errors.txt
@@ -1,0 +1,20 @@
+=== Metadata ===
+Effect version: 4.0.0
+
+/.src/catchDieToOrDie_preview.ts(9,3): warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catch` with an identity-forwarding `Effect.die` handler. effect(catchDieToOrDie)
+
+
+==== /.src/catchDieToOrDie_preview.ts (1 errors) ====
+    // @effect-v4
+    // @effect-diagnostics *:off
+    // @effect-diagnostics catchDieToOrDie:warning
+    import { Effect } from "effect"
+    
+    declare const readConfig: Effect.Effect<string, Error>
+    
+    export const program = readConfig.pipe(
+      Effect.catch((error) => Effect.die(error))
+      ~~~~~~~~~~~~
+!!! warning TS377115: `Effect.orDie` expresses escalating every typed failure into a defect more directly than `Effect.catch` with an identity-forwarding `Effect.die` handler. effect(catchDieToOrDie)
+    )
+
