+++ testdata/baselines/reference/effect-v4/catchDieToOrDie_preview.pipings.txt
@@ -1,4 +1,4 @@
-==== /.src/catchAllDieToOrDie_preview.ts (2 flows) ====
+==== /.src/catchDieToOrDie_preview.ts (2 flows) ====
 
 === Piping Flow ===
 Location: 8:23 - 10:2
